+++ R/predict.R
@@ -1,4 +1,4 @@
-#' Generate predictions from SLOPE models
+#' Generate Predictions from SLOPE Models
 #'
 #' Return predictions from models fit by [SLOPE()].
 #'
@@ -20,21 +20,23 @@
 #'   penalty path.
 #' @param sigma deprecated. Please use `alpha` instead.
 #'
-#' @seealso [stats::predict()], [stats::predict.glm()], [coef.SLOPE()]
+#' @seealso [stats::predict()], [stats::predict.glm()]
 #' @family SLOPE-methods
 #'
 #' @return Predictions from the model with scale determined by `type`.
 #'
+#' @export
+#'
 #' @examples
 #' fit <- with(mtcars, SLOPE(cbind(mpg, hp), vs, family = "binomial"))
 #' predict(fit, with(mtcars, cbind(mpg, hp)), type = "class")
-#' @export
 predict.SLOPE <- function(
   object,
   x,
   alpha = NULL,
   type = "link",
   simplify = TRUE,
+  exact = FALSE,
   sigma,
   ...
 ) {
@@ -53,13 +55,35 @@ predict.SLOPE <- function(
     alpha <- sigma
   }
 
-  beta <- getElement(object, "coefficients")
-  intercepts <- getElement(object, "intercepts")
+  # Get coefficients with intercepts using coef()
+  has_intercept <- getElement(object, "has_intercept")
+  coefs <- stats::coef(
+    object,
+    alpha = alpha,
+    exact = exact,
+    simplify = FALSE,
+    intercept = has_intercept,
+    ...
+  )
+
+  # Extract beta and intercepts from each matrix in the list
+  n_penalties <- length(coefs)
+  beta <- vector("list", n_penalties)
+  intercepts <- vector("list", n_penalties)
+
+  for (i in seq_len(n_penalties)) {
+    if (has_intercept) {
+      intercepts[[i]] <- coefs[[i]][1, ]
+      beta[[i]] <- coefs[[i]][-1, , drop = FALSE]
+    } else {
+      intercepts[[i]] <- rep(0, NCOL(coefs[[i]]))
+      beta[[i]] <- coefs[[i]]
+    }
+  }
 
   n <- NROW(x)
   p <- NROW(beta[[1L]])
   m <- NCOL(beta[[1L]])
-  n_penalties <- length(beta)
 
   stopifnot(p == NCOL(x))
 
@@ -73,7 +97,7 @@ predict.SLOPE <- function(
 
   for (i in seq_len(n_penalties)) {
     lp <- as.matrix(x %*% beta[[i]])
-    lin_pred[,, i] <- sweep(lp, 2, intercepts[[i]], "+")
+    lin_pred[,, i] <- sweep(lp, 2, intercepts[[i]], "+") # nolint
   }
 
   lin_pred