+++ inst/doc/introduction.R
@@ -64,7 +64,6 @@ for (i in seq_along(q)) {
   fdr[i] <- v / max(r, 1)
 }
 
-# Create the plot
 plot(
   q,
   fdr,
@@ -75,6 +74,5 @@ plot(
   las = 1
 )
 
-# Add horizontal line at 0.1
 abline(h = 0.1, lty = 3)